CRUD and JWT AUTH API is completed

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   //
   res.status(200).json(goals);
@@ -23,6 +24,7 @@ const setGoals = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -37,6 +39,20 @@ const updateGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // CHECK FOR USER IS EXIST OR NOT
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+
+  // CHECK TO MAKE SURE THE LOGGED IN USER MATHCES THE GOAL USER
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -55,6 +71,20 @@ const deleteGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // CHECK FOR USER IS EXIST OR NOT
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+
+  // CHECK TO MAKE SURE THE LOGGED IN USER MATHCES THE GOAL USER
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("user not authorized");
   }
 
   await Goal.findByIdAndRemove(req.params.id);
